automation/src/instagram/requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.8
influxdb==5.3.1
influxdb-client==1.45.0
instaloader==4.9.6
msgpack==1.0.8
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
reactivex==4.0.4
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
